third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-img-element/image-alternate-no-store.window-expected.txt

This is a testharness.js-based test.
[FAIL] Reassigning the same src value should not trigger an extra image fetch
  assert_equals: The image size should be the same expected 10 but got 100
Harness: the test ran to completion.
